Add files via upload
Our team developed a sophisticated model trained to analyze and predict various types of cancer based on their symptoms. Leveraging Flask, we established a robust web server that provides users with an intuitive platform for inputting symptoms and receiving accurate predictions regarding the likelihood of cancer.

This professional-grade solution integrates cutting-edge technology to ensure the reliability of predictions. The model, backed by comprehensive training in cancer pathology, enables users to make informed decisions about potential health concerns. The Flask-based web server not only ensures a seamless user experience but also adheres to industry standards for security and efficiency.

Our commitment to excellence in cancer prediction is reflected in the seamless synergy between advanced machine learning capabilities and user-friendly web interface. This tool is a testament to our dedication to leveraging technology for the betterment of healthcare, providing users with a reliable resource for early detection and intervention in cancer-related issues.